## What is the purpose of the change
The failure reason of AsyncWaitOperatorTest#testProcessingTimeRepeatedCompleteOrderedWithRetry maybe the slow execution occurred in internal executor service (which simulate external lookup of IllWrittenOddInputEmptyResultAsyncFunction), we improve the execution deadline from 1s to 10s so as to reduce the failure possiblilty.

## Brief change log
improve the execution deadline from 1s to 10s

## Verifying this change
AsyncWaitOperatorTest#testProcessingTimeRepeatedCompleteOrderedWithRetry

## Does this pull request potentially affect one of the following parts:
  - Dependencies (does it add or upgrade a dependency): (no)
  - The public API, i.e., is any changed class annotated with @Public(Evolving): (no)
  - The serializers: (no )
  - The runtime per-record code paths (performance sensitive): (no)
  - Anything that affects deployment or recovery: JobManager (and its components), Checkpointing, Kubernetes/Yarn, ZooKeeper: (no)
  - The S3 file system connector: (no)

## Documentation
  - Does this pull request introduce a new feature? (no)